fix: 채용공고 화면 조회 API 에러 수정

+++ src/main/java/com/example/demo/src/employment/EmploymentProvider.java
@@ -22,7 +22,7 @@ public class EmploymentProvider {
         this.employmentDao = employmentDao;
     }
 
-    public GetEmploymentRes getEmployments(int userIdx, String country, String sort, int years1, int years2) throws BaseException {
+    public GetEmploymentRes getEmployments(Long userIdx, String country, String sort, int years1, int years2) throws BaseException {
         try {
             GetEmploymentRes getEmploymentRes = employmentDao.getEmployments(userIdx, country, sort, years1, years2);
             return getEmploymentRes;
